backend/pyproject.toml
[project]
name = "poc-authentication-system"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "3.12.*"
fastapi-users = {extras = ["sqlalchemy"], version = "^14.0.1"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.37"}
pydantic-settings = "^2.7.1"
fastapi = {extras = ["standard"], version = "^0.115.8"}
aiosqlite = "^0.21.0"


[tool.poetry.group.dev.dependencies]
taskipy = "^1.13.0"
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
ruff = "^0.6.3"
httpx = "^0.27.2"
factory-boy = "^3.3.1"
freezegun = "^1.5.1"
ipdb = "^0.13.13"
testcontainers = "^4.9.0"
ipython = "^8.32.0"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix && ruff format .'
run = 'fastapi dev poc_authentication_system/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=fast_zero -vv'
post_test = 'coverage html'
